# Add support for forking at the scanning step