--- a/src/Psalm/Codebase/Scanner.php
+++ b/src/Psalm/Codebase/Scanner.php
@@ -9,6 +9,57 @@ use Psalm\Provider\FileStorageProvider;
 use Psalm\Scanner\FileScanner;
 
 /**
+ * @psalm-type  IssueData = array{
+ *     severity: string,
+ *     line_from: int,
+ *     line_to: int,
+ *     type: string,
+ *     message: string,
+ *     file_name: string,
+ *     file_path: string,
+ *     snippet: string,
+ *     from: int,
+ *     to: int,
+ *     snippet_from: int,
+ *     snippet_to: int,
+ *     column_from: int,
+ *     column_to: int
+ * }
+ *
+ * @psalm-type  PoolData = array{
+ *     classlikes_data:array{
+ *         0:array<string, bool>,
+ *         1:array<string, bool>,
+ *         2:array<string, bool>,
+ *         3:array<string, bool>,
+ *         4:array<string, bool>,
+ *         5:array<string, bool>,
+ *         6:array<string, bool>,
+ *         7:array<string, \PhpParser\Node\Stmt\Trait_>,
+ *         8:array<string, \Psalm\Aliases>,
+ *         9:array<string, int>
+ *     },
+ *     scanner_data:array{
+ *         0:array<string, string>,
+ *         1:array<string, string>,
+ *         2:array<string, string>,
+ *         3:array<string, bool>,
+ *         4:array<string, bool>,
+ *         5:array<string, string>,
+ *         6:array<string, bool>,
+ *         7:array<string, bool>,
+ *         8:array<string, bool>
+ *     },
+ *     issues:array<int, IssueData>,
+ *     changed_members:array<string, array<string, bool>>,
+ *     unchanged_signature_members:array<string, array<string, bool>>,
+ *     diff_map:array<string, array<int, array{0:int, 1:int, 2:int, 3:int}>>,
+ *     classlike_storage:array<string, \Psalm\Storage\ClassLikeStorage>,
+ *     file_storage:array<string, \Psalm\Storage\FileStorage>
+ * }
+ */
+
+/**
  * @internal
  *
  * Contains methods that aid in the scanning of Psalm's codebase
@@ -94,6 +145,11 @@ class Scanner
      * @var FileReferenceProvider
      */
     private $file_reference_provider;
+
+    /**
+     * @var bool
+     */
+    private $is_forked = false;
 
     /**
      * @param bool $debug_output
@@ -249,13 +305,13 @@ class Scanner
     /**
      * @return bool
      */
-    public function scanFiles(ClassLikes $classlikes)
+    public function scanFiles(ClassLikes $classlikes, int $pool_size = 1)
     {
         $has_changes = false;
 
         while ($this->files_to_scan || $this->classes_to_scan) {
             if ($this->files_to_scan) {
-                if ($this->scanFilePaths()) {
+                if ($this->scanFilePaths($pool_size)) {
                     $has_changes = true;
                 }
             } else {
@@ -266,7 +322,7 @@ class Scanner
         return $has_changes;
     }
 
-    private function scanFilePaths() : bool
+    private function scanFilePaths(int $pool_size) : bool
     {
         $filetype_scanners = $this->config->getFiletypeScanners();
         $files_to_scan = array_filter(
@@ -279,12 +335,108 @@ class Scanner
 
         $this->files_to_scan = [];
 
-        foreach ($files_to_scan as $file_path) {
-            $this->scanFile(
-                $file_path,
-                $filetype_scanners,
-                isset($this->files_to_deep_scan[$file_path])
+        $files_to_deep_scan = $this->files_to_deep_scan;
+
+        $scanner_worker =
+            /**
+             * @param int $_
+             * @param string $file_path
+             *
+             * @return void
+             */
+            function ($_, $file_path) use ($filetype_scanners, $files_to_deep_scan) {
+                $this->scanFile(
+                    $file_path,
+                    $filetype_scanners,
+                    isset($files_to_deep_scan[$file_path])
+                );
+            };
+
+        if (!$this->is_forked && $pool_size > 1 && count($files_to_scan) > 512) {
+            $pool_size = ceil(min($pool_size, count($files_to_scan) / 256));
+        } else {
+            $pool_size = 1;
+        }
+
+        if ($pool_size > 1) {
+            $process_file_paths = [];
+
+            $i = 0;
+
+            foreach ($files_to_scan as $file_path) {
+                $process_file_paths[$i % $pool_size][] = $file_path;
+                ++$i;
+            }
+
+            // Run scanning one file at a time, splitting the set of
+            // files up among a given number of child processes.
+            $pool = new \Psalm\Fork\Pool(
+                $process_file_paths,
+                /** @return void */
+                function () {
+                    $project_checker = \Psalm\Checker\ProjectChecker::getInstance();
+                    $statements_provider = $project_checker->codebase->statements_provider;
+
+                    $project_checker->codebase->scanner->isForked();
+                    $project_checker->codebase->file_storage_provider->deleteAll();
+                    $project_checker->codebase->classlike_storage_provider->deleteAll();
+
+                    $statements_provider->resetDiffs();
+                },
+                $scanner_worker,
+                /**
+                 * @return PoolData
+                */
+                function () {
+                    $project_checker = \Psalm\Checker\ProjectChecker::getInstance();
+                    $statements_provider = $project_checker->codebase->statements_provider;
+
+                    return [
+                        'classlikes_data' => $project_checker->codebase->classlikes->getThreadData(),
+                        'scanner_data' => $project_checker->codebase->scanner->getThreadData(),
+                        'issues' => \Psalm\IssueBuffer::getIssuesData(),
+                        'changed_members' => $statements_provider->getChangedMembers(),
+                        'unchanged_signature_members' => $statements_provider->getUnchangedSignatureMembers(),
+                        'diff_map' => $statements_provider->getDiffMap(),
+                        'classlike_storage' => $project_checker->classlike_storage_provider->getAll(),
+                        'file_storage' => $project_checker->file_storage_provider->getAll(),
+                    ];
+                }
             );
+
+            // Wait for all tasks to complete and collect the results.
+            /**
+             * @var array<int, PoolData>
+             */
+            $forked_pool_data = $pool->wait();
+
+            foreach ($forked_pool_data as $pool_data) {
+                \Psalm\IssueBuffer::addIssues($pool_data['issues']);
+
+                $this->codebase->statements_provider->addChangedMembers(
+                    $pool_data['changed_members']
+                );
+                $this->codebase->statements_provider->addUnchangedSignatureMembers(
+                    $pool_data['unchanged_signature_members']
+                );
+                $this->codebase->statements_provider->addDiffMap(
+                    $pool_data['diff_map']
+                );
+
+                $this->codebase->file_storage_provider->addMore($pool_data['file_storage']);
+                $this->codebase->classlike_storage_provider->addMore($pool_data['classlike_storage']);
+
+                $this->codebase->classlikes->addThreadData($pool_data['classlikes_data']);
+
+                $this->addThreadData($pool_data['scanner_data']);
+            }
+        } else {
+            $i = 0;
+
+            foreach ($files_to_scan as $file_path => $_) {
+                $scanner_worker($i, $file_path);
+                ++$i;
+            }
         }
 
         return (bool) $files_to_scan;
@@ -542,5 +694,84 @@ class Scanner
         }
 
         return true;
+    }
+
+    /**
+     * @return array{
+     *     0: array<string, string>,
+     *     1: array<string, string>,
+     *     2: array<string, string>,
+     *     3: array<string, bool>,
+     *     4: array<string, bool>,
+     *     5: array<string, string>,
+     *     6: array<string, bool>,
+     *     7: array<string, bool>,
+     *     8: array<string, bool>
+     * }
+     */
+    public function getThreadData()
+    {
+        return [
+            $this->files_to_scan,
+            $this->files_to_deep_scan,
+            $this->classes_to_scan,
+            $this->classes_to_deep_scan,
+            $this->store_scan_failure,
+            $this->classlike_files,
+            $this->deep_scanned_classlike_files,
+            $this->scanned_files,
+            $this->reflected_classlikes_lc
+        ];
+    }
+
+    /**
+     * @param array{
+     *     0: array<string, string>,
+     *     1: array<string, string>,
+     *     2: array<string, string>,
+     *     3: array<string, bool>,
+     *     4: array<string, bool>,
+     *     5: array<string, string>,
+     *     6: array<string, bool>,
+     *     7: array<string, bool>,
+     *     8: array<string, bool>
+     * } $thread_data
+     *
+     * @return void
+     */
+    public function addThreadData(array $thread_data)
+    {
+        list(
+            $files_to_scan,
+            $files_to_deep_scan,
+            $classes_to_scan,
+            $classes_to_deep_scan,
+            $store_scan_failure,
+            $classlike_files,
+            $deep_scanned_classlike_files,
+            $scanned_files,
+            $reflected_classlikes_lc
+        ) = $thread_data;
+
+        $this->files_to_scan = array_merge($files_to_scan, $this->files_to_scan);
+        $this->files_to_deep_scan = array_merge($files_to_deep_scan, $this->files_to_deep_scan);
+        $this->classes_to_scan = array_merge($classes_to_scan, $this->classes_to_scan);
+        $this->classes_to_deep_scan = array_merge($classes_to_deep_scan, $this->classes_to_deep_scan);
+        $this->store_scan_failure = array_merge($store_scan_failure, $this->store_scan_failure);
+        $this->classlike_files = array_merge($classlike_files, $this->classlike_files);
+        $this->deep_scanned_classlike_files = array_merge(
+            $deep_scanned_classlike_files,
+            $this->deep_scanned_classlike_files
+        );
+        $this->scanned_files = array_merge($scanned_files, $this->scanned_files);
+        $this->reflected_classlikes_lc = array_merge($reflected_classlikes_lc, $this->reflected_classlikes_lc);
+    }
+
+    /**
+     * @return void
+     */
+    public function isForked()
+    {
+        $this->is_forked = true;
     }
 }