--- a/src/Psalm/Codebase/Scanner.php
+++ b/src/Psalm/Codebase/Scanner.php
@@ -335,6 +335,10 @@ class Scanner
 
         $this->files_to_scan = [];
 
+        if (!$files_to_scan) {
+            return false;
+        }
+
         $files_to_deep_scan = $this->files_to_deep_scan;
 
         $scanner_worker =
@@ -439,7 +443,7 @@ class Scanner
             }
         }
 
-        return (bool) $files_to_scan;
+        return true;
     }
 
     /**