--- a/src/Psalm/Provider/FileStorageProvider.php
+++ b/src/Psalm/Provider/FileStorageProvider.php
@@ -87,6 +87,15 @@ class FileStorageProvider
     }
 
     /**
+     * @param array<string, FileStorage> $more
+     * @return void
+     */
+    public function addMore(array $more)
+    {
+        self::$storage = array_merge($more, self::$storage);
+    }
+
+    /**
      * @param  string $file_path
      *
      * @return FileStorage