This makes Psalm's scanning step faster by scanning files (parsing if necessary and crawling the AST to find relevant info) in separate processes.

The main aim is to reduce cold-start times when running a language server (coming soon), but it also benefits cold start times and from-cache start times when running Psalm with the `--threads=X` option.

It'll use threads to scan if the number of files in any single batch exceeds 512. I did this to cut down on the number of forks, as each fork incurs a cost both at the creation of the thread and then aggregating the information afterwards.

Scanning Vimeo's codebase with 10 threads (analysis time not included)
 - Before: **144s** cold, **28s** from cache
 - After: **53s** cold, **19s** from cache